Handle case were there's no match
It seems that sometimes when something goes wrong with Sublime Linter the match value may be falsey. Looking at linter.py, it checks match and may return it and a lot of None values if the it is not truthy. See:
https://github.com/SublimeLinter/SublimeLinter3/blob/master/lint/linter.py

I noticed this issue because match is None if an error occurs earlier in SublimeLinter.

I propose imitating the super function. An alternative would be to check the results of the call to super().split_match and then to just return the results if match or col are None.